SolarNasih_Deploiement_Complet/requirements_unified.txt
# ===== SOLAR NASIH - Requirements Unifiés =====
# Ce fichier contient toutes les dépendances pour SMA + RAG

# ===== CORE WEB FRAMEWORK =====
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.2

# ===== DATA VALIDATION =====
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== AI & LANGUAGE MODELS =====
langchain==0.0.350
sentence-transformers==2.2.2
transformers==4.36.0
google-generativeai==0.3.2
openai==1.3.0
anthropic==0.7.0
tavily-python==0.3.0

# ===== DEEP LEARNING =====
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# ===== VECTOR DATABASE =====
qdrant-client==1.6.0

# ===== IMAGE PROCESSING =====
Pillow>=10.0.0,<11.0.0
opencv-python>=4.8.0.74,<5.0.0

# ===== AUDIO PROCESSING =====
openai-whisper==20231117
librosa==0.10.1
speechrecognition==3.10.0
gtts==2.4.0

# ===== OCR =====
pytesseract==0.3.10

# ===== WEB & FILE HANDLING =====
streamlit==1.28.0
aiofiles==23.2.1

# ===== DATABASE & CACHE =====
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# ===== SCIENTIFIC COMPUTING =====
numpy>=1.24.0,<2.0.0
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0

# ===== UTILITIES =====
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
PyYAML==6.0.1
markdown==3.5.1
beautifulsoup4==4.12.2

# ===== DOCUMENT PROCESSING =====
python-docx==1.1.0
openpyxl==3.1.2
fpdf2==2.7.6
PyPDF2==3.0.1
pdf2image==1.16.3

# ===== AUDIO/VIDEO PROCESSING =====
ffmpeg-python==0.2.0

# ===== TESTING =====
pytest==7.4.3
pytest-asyncio==0.21.1

# ===== CLIP (Computer Vision) =====
git+https://github.com/openai/CLIP.git

# ===== WINDOWS COMPATIBILITY =====
pywin32==306; sys_platform == "win32"
